test(util): add util method to read TSP files
